% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CEdata}
\alias{CEdata}
\title{CEdata}
\format{
A data frame with 5571 rows and 7 variables:
\describe{
  \item{Age}{Age of reference person}
  \item{Urban}{Is this conumer unit located in an urban or rural area}
  \item{Tenrue}{Housing tenure}
  \item{Educ}{Education of reference person}
  \item{Expenditure}{Total expenditure last quarter}
  \item{Marital}{Marital status of reference person}
  \item{Income}{Total amount of family income before taxes in the last 12 months}
}
}
\usage{
CEdata
}
\description{
Sample of consumer expenditure survey
}
\keyword{datasets}
