% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.IdentificationRiskContinuousC}
\alias{.IdentificationRiskContinuousC}
\title{This function will compute the identification risk for a dataset with synthetic continuous and categorical variables.}
\usage{
.IdentificationRiskContinuousC(
  dataMatrix,
  rows,
  cols,
  syndataMatrices,
  num,
  knowncols,
  numKnown,
  syncols,
  numSyn,
  radius,
  percentage,
  categoricalVector
)
}
\arguments{
\item{radius}{radius to compare with for continous variables. Radius is either percentage (default) or fixed.
Radius can be the same for all continuous variables or specific to each. To specify for each use a vector, with
the radii ordered in the same order those columns appear in the dataset.}

\item{percentage}{true for a percentage radius, false for a constant radius}

\item{categoricalVector}{Boolean vector corresponding to the number of columns in the data, true means that column is categorical.}

\item{origdata}{dataframe of the origonal data}

\item{syndata}{list of the different synthetic dataframes}

\item{known}{vector of the names of the columns in the dataset assumed to be known}

\item{syn}{vector of the names of the columns in the dataset that are synthetic}
}
\description{
This function will compute the identification risk for a dataset with synthetic continuous and categorical variables.
}
