% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationRiskContinuous.R
\name{IdentificationRisk}
\alias{IdentificationRisk}
\title{An Identification Risk Function}
\usage{
IdentificationRisk(
  origdata,
  syndata,
  known,
  syn,
  r,
  percentage = TRUE,
  euclideanDist = FALSE
)
}
\arguments{
\item{origdata}{dataframe of the original data}

\item{syndata}{list of the different synthetic dataframes}

\item{known}{vector of the names of the columns in the dataset assumed to be known}

\item{syn}{vector of the names of the columns in the dataset that are synthetic}

\item{r}{radius to compare with for continuous variables. Radius is either percentage (default) or fixed.
Radius can be the same for all continuous variables or specific to each. To specify for each use a vector, with
the radii ordered in the same order those columns appear in the dataset.}

\item{percentage}{true for a percentage radius, false for a constant radius}

\item{euclideanDist}{true for a euclidean distance radius, false otherwise}
}
\description{
This function will compute the identification risk for a dataset with synthetic categorical variables.
This function assumes categorical variables will be as factors.
}
