% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationRiskCategorical.R
\name{IdentificationRiskCategorical}
\alias{IdentificationRiskCategorical}
\title{An Identification Risk Function}
\usage{
IdentificationRiskCategorical(origdata, syndata, known, syn)
}
\arguments{
\item{origdata}{dataframe of the origonal data}

\item{syndata}{list of the different synthetic dataframes}

\item{known}{vector of the names of the columns in the dataset assumed to be known}

\item{syn}{vector of the names of the columns in the dataset that are synthetic}
}
\description{
This function will compute the identification risk for a dataset with synthetic categorical variables.
}
