% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdentificationRiskContinuous.R
\name{IdentificationRiskContinuous}
\alias{IdentificationRiskContinuous}
\title{An Identification Risk Function}
\usage{
IdentificationRiskContinuous(
  origdata,
  syndata,
  known,
  syn,
  r,
  percentage = TRUE
)
}
\arguments{
\item{origdata}{dataframe of the origonal data}

\item{syndata}{list of the different synthetic dataframes}

\item{known}{vector of the names of the columns in the dataset assumed to be known}

\item{syn}{vector of the names of the columns in the dataset that are synthetic}

\item{r}{radius to compare with for continous variables. Radius is either percentage (default) or fixed}

\item{percentage}{true for a percentage radius, false for a constant radius}
}
\description{
This function will compute the identification risk for a dataset with synthetic categorical variables.
}
