% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{IdentificationRisk}
\alias{IdentificationRisk}
\title{This function will compute the identification risk for a dataset with synthetic categorical variables.}
\usage{
IdentificationRisk(
  dataMatrix,
  rows,
  cols,
  syndataMatrices,
  num,
  knowncols,
  numKnown,
  syncols,
  numSyn
)
}
\arguments{
\item{origdata}{dataframe of the origonal data}

\item{syndata}{list of the different synthetic dataframes}

\item{known}{vector of the names of the columns in the dataset assumed to be known}

\item{syn}{vector of the names of the columns in the dataset that are synthetic}
}
\description{
This function will compute the identification risk for a dataset with synthetic categorical variables.
}
